--- src/main/java/de/szut/lf8_project/domain/Project/Customer.java
@@ -1,4 +1,0 @@
-package de.szut.lf8_project.domain.Project;
-
-public class Customer {
-}
