--- src/main/java/de/szut/lf8_project/domain/Project/ProjectID.java
@@ -1,10 +1,0 @@
-package de.szut.lf8_project.domain.Project;
-
-import de.szut.lf8_project.common.ValueType;
-
-public class ProjectID extends ValueType<Long> {
-
-    ProjectID(Long value) {
-        super(value);
-    }
-}
