+++ src/main/java/de/szut/lf8_project/domain/customer/Customer.java
@@ -1,0 +1,8 @@
+package de.szut.lf8_project.domain.customer;
+
+import javax.validation.constraints.NotNull;
+
+public record Customer(
+        @NotNull CustomerId customerId
+        ) {
+}
