+++ src/main/java/de/szut/lf8_project/domain/employee/FirstName.java
@@ -1,4 +1,4 @@
-package de.szut.lf8_project.domain.Employee;
+package de.szut.lf8_project.domain.employee;
 
 import de.szut.lf8_project.common.ValueType;
