+++ src/main/java/de/szut/lf8_project/domain/project/ProjectId.java
@@ -1,0 +1,9 @@
+package de.szut.lf8_project.domain.project;
+
+import de.szut.lf8_project.common.ValueType;
+
+public class ProjectId extends ValueType<Long> {
+    public ProjectId(Long value) {
+        super(value);
+    }
+}
