+++ src/main/java/de/szut/lf8_project/domain/project/ProjectName.java
@@ -1,10 +1,9 @@
-package de.szut.lf8_project.domain.Project;
+package de.szut.lf8_project.domain.project;
 
 import de.szut.lf8_project.common.ValueType;
 
 public class ProjectName extends ValueType<String> {
-
-    ProjectName(String value) {
+    public ProjectName(String value) {
         super(value);
     }
 }
